+++ Cargo.toml
@@ -9,6 +9,7 @@ anyhow = "1.0.32"
 serde_json = "1.0.57"
 thiserror = "1.0"
 nom = "5.1.2"
+lazy_static = "1.4.0"
 
 [build-dependencies]
 yaml-rust = "0.4"
