+++ src/lib.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 mod errors;
+mod builtins;
 mod interpreter;
 mod render;
 mod value;
